Refatora backend: modulariza rotas, adiciona dotenv e tratamento de erros
