+++ backend/routes/productRoutes.js
@@ -1,0 +1,80 @@
+const express = require('express');
+const router = express.Router();
+const multer = require('multer');
+
+// Multer setup for file uploads
+const upload = multer({ dest: 'uploads/' });
+
+// 1. Quantum Data Ingestion (File Upload)
+router.post('/upload', upload.single('data'), (req, res) => {
+  if (!req.file) {
+    return res.status(400).send('No file uploaded.');
+  }
+  console.log('File received:', req.file.originalname);
+  // In a real application, process the file here (e.g., store, analyze, pass to quantum backend)
+  res.json({ message: 'File uploaded successfully', filename: req.file.originalname, size: req.file.size });
+});
+
+// 3. Quantum Job Execution (General)
+router.post('/ia', (req, res) => {
+  const { module, prompt } = req.body;
+  console.log(`Quantum job request: Module - ${module}, Prompt - ${prompt}`);
+  // In a real application: interact with quantum backend/API
+  res.json({ result: `Quantum job for '${prompt}' processed by '${module}' (placeholder)` });
+});
+
+// --- Sector-Specific Placeholder Endpoints ---
+
+// QuantumPredict Pro Endpoints
+router.post('/predict/aerospace-trajectory', (req, res) => {
+  const { data } = req.body;
+  console.log('Predicting aerospace trajectory with quantum algorithms:', data);
+  res.json({ prediction: 'Optimized trajectory calculated (placeholder)' });
+});
+
+router.post('/predict/automotive-maintenance', (req, res) => {
+  const { sensorData } = req.body;
+  console.log('Predicting automotive maintenance needs with QML:', sensorData);
+  res.json({ prediction: 'Next maintenance due in X km (placeholder)' });
+});
+
+// QuantumRisk Guardian Endpoints
+router.post('/risk/maritime-supply-chain', (req, res) => {
+  const { manifest, route } = req.body;
+  console.log('Assessing maritime supply chain risk with quantum algorithms:', manifest, route);
+  res.json({ risk_assessment: 'Low risk, optimized route (placeholder)' });
+});
+
+router.post('/risk/space-cybersecurity', (req, res) => {
+  const { systemId } = req.body;
+  console.log('Assessing space cybersecurity posture with PQC:', systemId);
+  res.json({ security_status: 'Post-quantum secure (placeholder)' });
+});
+
+// QuantumEco AI Endpoints
+router.post('/eco/automotive-charging-optimization', (req, res) => {
+  const { vehicleId, location, time } = req.body;
+  console.log('Optimizing EV charging for:', vehicleId, location, time);
+  res.json({ optimal_charging_schedule: 'Charge from 2 AM to 4 AM (placeholder)' });
+});
+
+router.post('/eco/aero-material-design', (req, res) => {
+  const { materialSpecs } = req.body;
+  console.log('Designing aerospace material with quantum simulation:', materialSpecs);
+  res.json({ optimized_material_composition: 'New lightweight alloy (placeholder)' });
+});
+
+// QuantumSense Connect Endpoints
+router.post('/sense/maritime-structural-integrity', (req, res) => {
+  const { sensorReadings } = req.body;
+  console.log('Analyzing maritime structural integrity from quantum sensors:', sensorReadings);
+  res.json({ integrity_report: 'No anomalies detected (placeholder)' });
+});
+
+router.post('/sense/space-navigation-precision', (req, res) => {
+  const { navData } = req.body;
+  console.log('Enhancing space navigation precision with quantum sensor data:', navData);
+  res.json({ enhanced_position: 'Lat: X, Lon: Y, Alt: Z (placeholder)' });
+});
+
+module.exports = router;
