+++ backend/server.js
@@ -1,125 +1,26 @@
+require('dotenv').config(); // Load environment variables
 const express = require('express');
 const cors = require('cors');
-const multer = require('multer');
 const app = express();
-const port = 3000;
+const port = process.env.PORT || 3000; // Use port from .env or default to 3000
+
+// Import routes
+const authRoutes = require('./routes/authRoutes');
+const productRoutes = require('./routes/productRoutes');
 
 // Middleware
 app.use(cors());
 app.use(express.json()); // For parsing application/json
 app.use(express.urlencoded({ extended: true })); // For parsing application/x-www-form-urlencoded
 
-// Multer setup for file uploads
-const upload = multer({ dest: 'uploads/' });
+// Use routes
+app.use('/api', authRoutes);
+app.use('/api', productRoutes);
 
-// --- Placeholder API Endpoints ---
-
-// 1. Quantum Data Ingestion (File Upload)
-app.post('/api/upload', upload.single('data'), (req, res) => {
-  if (!req.file) {
-    return res.status(400).send('No file uploaded.');
-  }
-  console.log('File received:', req.file.originalname);
-  // In a real application, process the file here (e.g., store, analyze, pass to quantum backend)
-  res.json({ message: 'File uploaded successfully', filename: req.file.originalname, size: req.file.size });
-});
-
-// 2. Authentication Endpoints
-// Placeholder for user registration
-app.post('/api/register', (req, res) => {
-  const { username, password } = req.body;
-  console.log(`Register attempt: ${username}`);
-  // In a real app: hash password, save to DB, generate token
-  res.json({ msg: `User ${username} registered (placeholder)` });
-});
-
-// Placeholder for user login
-app.post('/api/login', (req, res) => {
-  const { username, password } = req.body;
-  console.log(`Login attempt: ${username}`);
-  // In a real app: verify credentials, generate JWT
-  if (username === 'test' && password === 'password') {
-    res.json({ access_token: 'mock_jwt_token', msg: 'Login successful (placeholder)' });
-  } else {
-    res.status(401).json({ msg: 'Invalid credentials (placeholder)' });
-  }
-});
-
-// Placeholder for fetching user data (requires token)
-app.get('/api/users/me', (req, res) => {
-  const authHeader = req.headers.authorization;
-  if (!authHeader || !authHeader.startsWith('Bearer ')) {
-    return res.status(401).send('Authorization token required.');
-  }
-  const token = authHeader.split(' ')[1];
-  // In a real app: validate token, fetch user from DB
-  if (token === 'mock_jwt_token') {
-    res.json({ username: 'testuser', email: 'test@example.com', roles: ['user'] });
-  } else {
-    res.status(403).send('Invalid token.');
-  }
-});
-
-// 3. Quantum Job Execution (General)
-app.post('/api/ia', (req, res) => {
-  const { module, prompt } = req.body;
-  console.log(`Quantum job request: Module - ${module}, Prompt - ${prompt}`);
-  // In a real application: interact with quantum backend/API
-  res.json({ result: `Quantum job for '${prompt}' processed by '${module}' (placeholder)` });
-});
-
-// --- New Sector-Specific Placeholder Endpoints ---
-
-// QuantumPredict Pro Endpoints
-app.post('/api/predict/aerospace-trajectory', (req, res) => {
-  const { data } = req.body;
-  console.log('Predicting aerospace trajectory with quantum algorithms:', data);
-  res.json({ prediction: 'Optimized trajectory calculated (placeholder)' });
-});
-
-app.post('/api/predict/automotive-maintenance', (req, res) => {
-  const { sensorData } = req.body;
-  console.log('Predicting automotive maintenance needs with QML:', sensorData);
-  res.json({ prediction: 'Next maintenance due in X km (placeholder)' });
-});
-
-// QuantumRisk Guardian Endpoints
-app.post('/api/risk/maritime-supply-chain', (req, res) => {
-  const { manifest, route } = req.body;
-  console.log('Assessing maritime supply chain risk with quantum algorithms:', manifest, route);
-  res.json({ risk_assessment: 'Low risk, optimized route (placeholder)' });
-});
-
-app.post('/api/risk/space-cybersecurity', (req, res) => {
-  const { systemId } = req.body;
-  console.log('Assessing space cybersecurity posture with PQC:', systemId);
-  res.json({ security_status: 'Post-quantum secure (placeholder)' });
-});
-
-// QuantumEco AI Endpoints
-app.post('/api/eco/automotive-charging-optimization', (req, res) => {
-  const { vehicleId, location, time } = req.body;
-  console.log('Optimizing EV charging for:', vehicleId, location, time);
-  res.json({ optimal_charging_schedule: 'Charge from 2 AM to 4 AM (placeholder)' });
-});
-
-app.post('/api/eco/aero-material-design', (req, res) => {
-  const { materialSpecs } = req.body;
-  console.log('Designing aerospace material with quantum simulation:', materialSpecs);
-  res.json({ optimized_material_composition: 'New lightweight alloy (placeholder)' });
-});
-
-// QuantumSense Connect Endpoints
-app.post('/api/sense/maritime-structural-integrity', (req, res) => {
-  const { sensorReadings } = req.body;
-  console.log('Analyzing maritime structural integrity from quantum sensors:', sensorReadings);
-  res.json({ integrity_report: 'No anomalies detected (placeholder)' });
-});
-
-app.post('/api/sense/space-navigation-precision', (req, res) => {
-  const { navData } = req.body;
-  console.log('Enhancing space navigation precision with quantum sensor data:', navData);
-  res.json({ enhanced_position: 'Lat: X, Lon: Y, Alt: Z (placeholder)' });
+// Basic Error Handling Middleware
+app.use((err, req, res, next) => {
+  console.error(err.stack);
+  res.status(500).send('Something broke!');
 });
 
 // Start the server
